yaml: remove drm-color-bar from build
We do not need to blacklist some modules.
Also current state of the recipe doesn't work
as expected with kirkstone, due to conflict between
systemd and sysVinit. `inherit systemd` results in
erasing of ${D}/etc/init.d folder with required script.
According to the doc `This will enable systemd and
remove sysvinit components from the image.`

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>