optee: fix machine suffix handling
The previous install rule hardcoded the "spider" suffix in the
"tee.srec" filename, regardless of the selected machine
("spider" or "s4sk").

This caused an error when running the "boot_artifacts" for the "s4sk"
machine, because it expected "tee-s4sk.srec", but really
"tee-spider.srec" was created.

Now the machine-suffix is added correctly.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>