Create symbolic link "tee.srec" in the deploy directory
This patch creates a symlink "tee.srec" in the deploy directory
that always points to the actual file "tee-${MACHINE}.srec",
depending on the selected board.

The purpose of this symlink is to simplify the flashing process using
"rcar_flash".
With this change, "rcar_flash" can refer to the generic "tee.srec"
symlink, without specifying the machine suffix.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>